cart notification / line item script conflict fix
from Shopify support:

This will allow the cart notification to show and the cart count to be updated. The caveat here is if line item properties were to be used within your theme, as we're now using the variant ID to render the item information rather than the individual cart item key, properties would not display correctly within the cart notification, but as you don't appear to be using them, this should resolve the issue for you. I can't give a timeline on when this might be resolved entirely unfortunately.